requirements.txt
google-cloud-bigquery
mariadb
python-dotenv
requests
